split journal landing page aspect out into a separate module for dev
