+++ dspace/modules/xmlui-journal-landing/xmlui-journal-landing-webapp/pom.xml
@@ -1,0 +1,123 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>org.dspace.modules</groupId>
+    <artifactId>xmlui-journal-landing-webapp</artifactId>
+	<packaging>war</packaging>
+    <name>Dryad Journal Landing Pages :: Web Application</name>
+    <description>Dryad Journal Landing Pages</description>
+
+	<parent>
+		<groupId>org.dspace.modules</groupId>
+	    <artifactId>xmlui-journal-landing</artifactId>
+		<version>1.7.3-SNAPSHOT</version>
+	</parent>
+
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-war-plugin</artifactId>
+				<configuration>
+					<archiveClasses>true</archiveClasses>
+                    <failOnMissingWebXml>false</failOnMissingWebXml>
+                    <attachClasses>true</attachClasses>
+                    <packagingExcludes>WEB-INF/lib/*.jar</packagingExcludes>
+                    <warSourceExcludes>WEB-INF/lib/*.jar</warSourceExcludes>
+					<webResources>
+						<resource>
+							<filtering>true</filtering>
+							<directory>${basedir}/src/main/webapp</directory>
+							<includes>
+								<include>WEB-INF/web.xml</include>
+							</includes>
+						</resource>
+					</webResources>
+				</configuration>
+				<executions>
+					<execution>
+						<phase>prepare-package</phase>
+					</execution>
+				</executions>
+			</plugin>
+		</plugins>
+	</build>
+
+	<profiles>
+		<!--
+			when activated a dspace.config configuration file location passed on
+			the commandline (-Ddspace.config=...) can be passed through to be
+			used as a filter source by projects for tasks such as updating the
+			${dspace.dir} in web.xml etc.
+		-->
+		<profile>
+			<activation>
+				<property>
+					<name>dspace.config</name>
+				</property>
+			</activation>
+			<build>
+				<filters>
+					<filter>${dspace.config}</filter>
+				</filters>
+			</build>
+		</profile>
+	    
+	    <profile>
+	        <id>dev</id>
+	        <dependencies>
+	            <dependency>
+	                <groupId>org.dspace</groupId>
+	                <artifactId>dspace-api</artifactId>
+	                <version>${pom.version}</version>
+	            </dependency>
+	            <dependency>
+	                <groupId>org.dspace</groupId>
+	                <artifactId>dspace-xmlui-api</artifactId>
+	                <version>${pom.version}</version>
+	            </dependency>
+	            <dependency>
+	                <groupId>org.dspace</groupId>
+	                <artifactId>dspace-xmlui-wing</artifactId>
+	                <version>${pom.version}</version>
+	            </dependency>
+	            <dependency>
+	                <groupId>org.dspace</groupId>
+	                <artifactId>dspace-xmlui-webapp</artifactId>
+	                <type>war</type>
+	                <version>${pom.version}</version>
+	            </dependency>
+	            <dependency>
+	                <groupId>org.apache.avalon.framework</groupId>
+	                <artifactId>avalon-framework-api</artifactId>
+	                <version>4.3.1</version>
+	            </dependency>
+	            <dependency>
+	                <groupId>org.apache.cocoon</groupId>
+	                <artifactId>cocoon-sitemap-api</artifactId>
+	                <version>1.0.0</version>
+	            </dependency>	            
+	            <dependency>
+	                <groupId>org.apache.cocoon</groupId>
+	                <artifactId>cocoon-sitemap-impl</artifactId>
+	                <version>1.0.0</version>
+	            </dependency>
+	            <dependency>
+	                <groupId>log4j</groupId>
+	                <artifactId>log4j</artifactId>
+	                <version>1.2.14</version>
+	            </dependency>
+	        </dependencies>
+	    </profile>
+	    
+	</profiles>
+
+	<dependencies>
+        <dependency>
+           <groupId>org.dspace.modules</groupId>
+           <artifactId>api-stats</artifactId>
+           <version>${pom.version}</version>
+        </dependency>
+	</dependencies>
+
+</project>
