+++ dspace/modules/xmlui-journal-landing/xmlui-journal-landing-webapp/src/main/java/org/dspace/app/xmlui/aspect/journal/landing/Const.java
