+++ dspace/modules/xmlui-journal-landing/xmlui-journal-landing-webapp/src/main/java/org/dspace/app/xmlui/aspect/journal/landing/UserGeography.java
@@ -46,6 +46,8 @@ public class UserGeography extends AbstractDSpaceTransformer {
         // Geographic breakdown of users
         // 
         // ------------------
+
+        // see com.maxmind.geoip
     }
 
 }
