+++ dspace/modules/xmlui-journal-landing/xmlui-journal-landing-webapp/src/main/resources/aspects/JournalLandingPage/il8n/messges.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0"?>
+<catalogue xml:lang="en" xmlns:i18n="http://apache.org/cocoon/i18n/2.1">
+
+    <!--
+         The format used by all keys is as follows
+
+         xmlui.<Aspect>.<Java Class>.<name>
+
+         There are a few exceptions to this nameing format,
+         1) Some general keys are in the xmlui.general namespace
+            because they are used very frequently.
+         2) Some general keys which are specific to a particular aspect
+            may be found at xmlui.<Aspect> without specifiying a
+            particular java class.
+         -->
+    <!--
+    <message key="xmlui.JournalLandingPage."></message>
+    -->
+
+</catalogue>
